+++ BankManagementSystem/Entity/Client.cs
@@ -13,11 +13,11 @@ namespace BankManagementSystem
         public string Lastname { get; set; }
         public string Address { get; set; }
         public string Email { get; set; }
-        public int PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
         public string DOB { get; set; }
         public string Nationality { get; set; }
         public string Gender { get; set; }
-        public int NID { get; set; }
+        public long NID { get; set; }
         public string Occupation { get; set; }
         public string ImageDir { get; set; }
         //private Account account;
